Fix bug that causes trap on Linux
This patch fixes a bug that causes a trap, SIGABRT, buffer overflow,
etc., when printf() is called after read_tuples_csv() executes. This
is caused by using feof() as a switch for the while loop. This is not
only bad practice, but it delivered information irrelevant to the
program and caused the program to crash at the next invocation of
printf(). Fix by instead switching the loop on the return value of
getline().