First STABLE commit
This commit marks the first stable commit. The method was changed to
use the GSL's finite difference method for computing the Jacobian,
which resulted in a successful run. In the future, surface_df() will
be changed to correctly provide the Jacobian matrix. In addition, the
vector of initial values was changed from {5.0} to {1.0}. Finally,
the formwat for the callback function was changed to provide for
legible output.