Add Valgrind output detailing memory leaks & remove comments fn
Add a file resulting from Valgrind which details the memory leaks
created by the program. Obviously, this is unacceptable. Measures
will be taken soon to prevent these from happening. Also add a fn
to remove comments from strings read from the .csv file. The fn
will skip the line if it contains only comment. Updates will be made
to this soon to make it more robust.